removed unused components, renamed "new-timeline" into "timeline"
