+++ tools/gestenrecorder/src/app/app.component.ts
@@ -3,7 +3,7 @@ import { RouterOutlet } from "@angular/router";
 import { TouchAreaComponent } from "./touch-area/touch-area.component";
 import { DropdownComponent } from "./dropdown/dropdown.component";
 import { PullupComponent } from "./pullup/pullup.component";
-import { NewTimelineComponent } from "./new-timeline/new-timeline.component";
+import { NewTimelineComponent } from "./timeline/timeline.component";
 import { GestureDataService } from "./service/gesture-data.service";
 import { GestureReplayService } from "./service/gesture-replay.service";
