--- tools/gestenrecorder/src/app/new-timeline/new-timeline.component.html
@@ -1,18 +1,0 @@
-<div class="timeline-container">
-    <div>
-        <!-- Bahn mit Segmenten -->
-        <!-- <div class="track-container">
-            <div class="track"></div>
-            <div class="segments-container">
-                <div *ngFor="let segment of segments; let i = index" [style.width.px]="segmentWidth ? segmentWidth : null" class="segment" (click)="updateHorizontalPosition(i)"></div>
-            </div>
-        </div> -->
-        <!-- Plotly-Graph -->
-        <div class="graph-container" id="graph-container">
-            <div class="overlay"></div>
-            <div class="vertical-line"></div>
-            <plotly-plot [data]="graph.data" [layout]="graph.layout"></plotly-plot>
-        </div>
-    </div>
-
-</div>
