--- tools/gestenrecorder/src/app/options/options.component.html
@@ -1,138 +1,0 @@
-<h3 class="options__heading">Server</h3>
-  <!-- server address -->
-    <div class="form-group">
-      <label for="serverConnection">Address</label>
-      <input type="text" class="form-control" id="serverConnection" name="serverConnection" placeholder="enter server address" [(ngModel)]="serverConnection" (change)="saveConfiguration()">
-      <div class="connection-indicator" #indicator></div>
-    </div>
-
-  <!-- send interval -->
-  <div class="row">
-    <div class="form-group col-sm-6">
-        <label for="sendInterval">Send Interval
-            <span class="fa fa-question-circle options__tooltip" data-toggle="tooltip" data-placement="top" title="interval in which tracking data is sent"></span>
-        </label>
-        <span class="input-unit-container">
-            <input type="number" class="form-control" id="sendInterval" name="sendInterval" [(ngModel)]="sendInterval" (change)="saveConfiguration()">
-            <span class="input-unit">ms</span>
-        </span>
-    </div>
-  </div>
-
-  <h3 class="options__heading">Touch Points</h3>
-
-  <div class="row">
-      <!-- amount -->
-      <div class="form-group col-sm-4">
-          <label for="amountTouchPoints">Max Amount
-              <span class="fa fa-question-circle options__tooltip" data-toggle="tooltip" data-placement="top" title="amount of touches tracked at the same time"></span>
-          </label>
-          <input type="number" min="1" class="form-control" id="amountTouchPoints" name="amountTouchPoints" [(ngModel)]="amountTouchPoints" (change)="saveConfiguration()">
-      </div>
-      <!-- min circle size -->
-      <div class="form-group col-sm-4">
-          <label for="circleSize">Min Size</label>
-          <input type="number" min="10" max="200" class="form-control" id="circleMinSize" name="circleMinSize" [(ngModel)]="circleSize.min" (change)="saveConfiguration()">
-      </div>
-      <!-- max circle size -->
-      <div class="form-group col-sm-4">
-          <label for="circleSize">Max Size</label>
-          <input type="number" min="10" max="400" class="form-control" id="circleMaxSize" name="circleMaxSize" [(ngModel)]="circleSize.max" (change)="saveConfiguration()">
-      </div>
-  </div>
-
-  <!-- viewport dimensions-->
-  <h3 class="options__heading">Viewport</h3>
-  <div class="row">
-      <div class="form-group col-sm-6">
-          <label for="viewPortWidth">Width</label>
-          <span class="input-unit-container">
-              <input type="number" class="form-control" id="viewPortWidth" name="viewPortWidth" [(ngModel)]="viewPort.width" (change)="saveConfiguration()">
-              <span class="input-unit">px</span>
-          </span>
-      </div>
-      <div class="form-group col-sm-6">
-          <label for="viewPortHeight">Height</label>
-          <span class="input-unit-container">
-              <input type="number" class="form-control" id="viewPortHeight" name="viewPortHeight" [(ngModel)]="viewPort.height" (change)="saveConfiguration()">
-              <span class="input-unit">px</span>
-          </span>
-      </div>
-  </div>
-
-  <!-- layers -->
-  <div class="form-group clearfix">
-      <label for="projectionLayers">Layers
-          <span class="fa fa-question-circle options__tooltip" data-toggle="tooltip" data-placement="top" title="number of depths layers, max. 20 each"></span>
-      </label>
-
-      <div class="row">
-          <div class="form-group col-sm-6">
-              <label for="viewPortWidth">Pull</label>&nbsp;
-              <span class="fas fa-arrow-up"></span>
-              <span class="input-unit-container">
-                  <input type="number" min="0" max="20" class="form-control" id="layersUp" name="layersUp" [(ngModel)]="layers.up" (change)="saveConfiguration()">
-              </span>
-          </div>
-          <!-- color / NOTE: to pass changes on color inputs immediately use "(input)=..." -->
-          <div class="form-group col-sm-6">
-              <label for="colorUp">Color</label>
-              <input type="color" class="form-control" id="colorUp" name="colorUp" value="#8ee6d0" [(ngModel)]="layers.colorUp" (input)="saveConfiguration()">
-          </div>
-      </div>
-      <div class="row">
-          <div class="form-group col-sm-6">
-              <label for="viewPortHeight">Push</label>&nbsp;
-              <span class="fas fa-arrow-down"></span>
-              <span class="input-unit-container">
-                  <input type="number" min="0" max="20" class="form-control" id="layersDown" name="layersDown" [(ngModel)]="layers.down" (change)="saveConfiguration()">
-              </span>
-          </div>
-          <!-- color / NOTE: to pass changes on color inputs immediately use "(input)=..." -->
-          <div class="form-group col-sm-6">
-              <label for="colorDown">Color</label>
-              <input type="color" class="form-control" id="colorDown" name="colorDown" value="#00e6d0" [(ngModel)]="layers.colorDown" (input)="saveConfiguration()">
-          </div>
-      </div>
-
-  </div>
-  <hr>
-
-  <h3 class="options__heading">Background</h3>
-
-  <!-- projection sources -->
-  <div class="form-group">
-      <div class="custom-control custom-radio custom-control-inline">
-          <input type="radio" id="backgroundType0" name="backgroundType" class="custom-control-input" [value]='0' [(ngModel)]='backgroundType' (change)="onBackgroundSelected()">
-          <label class="custom-control-label" for="backgroundType0">none</label>
-      </div>
-      <div class="custom-control custom-radio custom-control-inline">
-          <input type="radio" id="backgroundType1" name="backgroundType" class="custom-control-input" [value]='1' [(ngModel)]='backgroundType' (change)="onBackgroundSelected()">
-          <label class="custom-control-label" for="backgroundType1">select Image</label>
-      </div>
-      <div class="custom-control custom-radio custom-control-inline">
-          <input type="radio" id="backgroundType2" name="backgroundType" class="custom-control-input" [value]='2' [(ngModel)]='backgroundType' (change)="onBackgroundSelected()">
-          <label class="custom-control-label" for="backgroundType2">via URL</label>
-      </div>
-
-      <!-- (conditionally rendered) inputs for projection sources -->
-      <select class="custom-select" name="backgroundImage" [(ngModel)]="backgroundImage" *ngIf="backgroundType===1" (change)="onBackgroundSelected()">
-        <option *ngFor="let img of backgroundSources; let i = index" [value]="img.src">{{ img.name }}</option>
-      </select>
-      <div *ngIf="backgroundType===2">
-        <input type="text" class="form-control" name="backgroundUrl" [(ngModel)]='backgroundUrl' accept="image/*" (change)="onBackgroundSelected()">
-      </div>
-  </div>
-
-  <div class="options__absolute row">
-    <div class="form-group col-sm-6">
-        <button class="btn btn-secondary btn-block options__save-btn" (click)="restoreConfiguration()">
-            restore
-      </button>
-    </div>
-    <div class="form-group col-sm-6">
-        <button class="btn btn-primary btn-block options__save-btn" (click)="saveConfiguration()">
-            save
-      </button>
-    </div>
-  </div>
