--- tools/gestenrecorder/src/app/options/options.component.scss
@@ -1,119 +1,0 @@
-@import 'reflex';
-
-.connection-indicator {
-  position: absolute;
-  top: 50%;
-  right: 0;
-  transform: translateY(-50%);
-  width: 10px;
-  height: 10px;
-  border-radius: 100%;
-  border: 1px solid $color-grey-light-2;
-  background: $color-grey-light-2;
-}
-
-.connection-indicator.success {
-  background: $color-success;
-}
-
-.connection-indicator.error {
-  background: $color-error-light;
-}
-
-.form-group {
-  position: relative;
-  label {
-    font-size: 10px;
-  }
-
-  span {
-    font-size: 10px;
-  }
-}
-
-.options__absolute {
-  position: relative;
-  bottom: 0;
-  width: 100%;
-}
-
-.options__container {
-  position: absolute;
-  right: 0;
-  top: 0;
-  border-radius: 5px;
-  margin: 10px;
-  padding: 10px;
-  background: $color-black;
-  color: $color-grey-light-4;
-  height: calc(75% - 20px);
-  overflow-y: scroll;
-}
-
-.options__container::-webkit-scrollbar {
-  display: none;
-}
-
-.options__heading {
-  padding-top: 5px;
-  font-size: 12px;
-}
-
-.options__content {
-  position: relative;
-  padding: 0 10px 0 0;
-  height: 100%;
-  font-size: 0.8rem;
-}
-
-.options__content .row {
-  margin: 0;
-}
-
-.options__content .form-group {
-  position: relative;
-  padding: 5px;
-  padding-bottom: 0;
-  margin: 0;
-}
-
-.options__content .form-group .custom-radio {
-  margin: 0 10px 5px 0;
-}
-
-.options__content .form-group .custom-radio::after {
-  font-size: 0.8rem;
-}
-
-.options__content input[type="color"] {
-  display: block;
-  padding: 0;
-}
-
-.options__save-btn {
-  margin-top: 10px;
-}
-
-.options__save-btn.btn-secondary {
-  background-color: $color-grey-light;
-  color: $color-white;
-}
-
-.options__save-btn.btn-secondary:active {
-  background-color: $color-grey-light;
-  color: $color-white;
-}
-
-.options__save-btn.btn-primary {
-  background-color: $color-light-2_100;
-  color: $color-white;
-}
-
-.options__save-btn.btn-primary:active {
-  background-color: $color-light-2_100;
-  color: $color-white;
-}
-
-.form-group.row {
-  margin-top: 20px;
-}
