+++ tools/gestenrecorder/src/app/pullup/pullup.component.html
@@ -13,7 +13,7 @@
     <table class="pullup__table">
       <tr>
         <td>
-          <app-new-timeline></app-new-timeline>
+          <app-timeline></app-timeline>
         </td>
         <td>
           <app-track-component></app-track-component>
