+++ tools/gestenrecorder/src/app/pullup/pullup.component.scss
@@ -28,7 +28,7 @@
 // ##### pullup itself #####
 .pullup__area {
   position: fixed;
-  background-color: white;
+  background-color: $color-white;
   border: 1px solid $color-primary;
   border-radius: 4px;
   padding: 10px;
