--- tools/gestenrecorder/src/app/timeline/timeline-canvas/timeline-canvas.component.html
@@ -1,5 +1,0 @@
-<canvas
-    #timeline
-    id="timeline"
-    class="canvas"
-></canvas>
