--- tools/gestenrecorder/src/app/timeline/timeline-canvas/timeline-canvas.component.scss
@@ -1,16 +1,0 @@
-@import "../../../styles";
-
-:host {
-    height: 100%;
-    width: 100%;
-    position: relative;
-    display: inline-block;
-    overflow: scroll;
-    box-sizing: border-box;
-
-    &::-webkit-scrollbar {
-        display: none;
-    }
-}
-
-
