+++ tools/gestenrecorder/src/app/timeline/timeline.component.html
@@ -1,6 +1,18 @@
-<div class="timeline__container">
-    <div class="timeline__header">
-        <h2 class="timeline__title">Timeline</h2>
+<div class="timeline-container">
+    <div>
+        <!-- Bahn mit Segmenten -->
+        <!-- <div class="track-container">
+            <div class="track"></div>
+            <div class="segments-container">
+                <div *ngFor="let segment of segments; let i = index" [style.width.px]="segmentWidth ? segmentWidth : null" class="segment" (click)="updateHorizontalPosition(i)"></div>
+            </div>
+        </div> -->
+        <!-- Plotly-Graph -->
+        <div class="graph-container" id="graph-container">
+            <div class="overlay"></div>
+            <div class="vertical-line"></div>
+            <plotly-plot [data]="graph.data" [layout]="graph.layout"></plotly-plot>
+        </div>
     </div>
-    <app-timeline-canvas id="timeline-wrapper"></app-timeline-canvas>
+
 </div>
