+++ tools/gestenrecorder/src/app/timeline/timeline.component.scss
@@ -1,24 +1,66 @@
 @import 'reflex';
 
-.timeline {
-    &__container {
-        position: relative;
-        bottom: 0;
-        right: 0;
-        margin: 5px;
-        background-color: white;
-    }
+.timeline-container{
+    display: flex;
+}
+.timeline-button1,.timeline-button2,.timeline-button3{
+    width: 100%;
+}
 
-    &__header {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-bottom: 8px;
+.modebar-container {
+    display: none !important;
+}
 
-        .timeline__title {
-            margin: 0;
-            font-size: 16px;
-            color: $color-primary;
-        }
-    }
+.track-container {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    position: relative;
+}
+
+.track {
+    flex: 1;
+    height: 20px;
+    background-color: $color-white;
+    border: 1px solid $color-grey-light;
+}
+
+.segments-container {
+    flex: 1;
+    display: flex;
+    justify-content: space-between;
+    position: fixed;
+}
+
+.segment {
+    height: 20px;
+    background-color: $color-white;
+    cursor: pointer;
+    border-left: 1px solid $color-primary;
+    border-right: 1px solid $color-grey-light;
+    border-bottom: 1px solid $color-grey-light;
+}
+
+.graph-container {
+    position: relative;
+}
+
+.overlay {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba($color-black, 0.01);
+    z-index: 10;
+}
+
+.vertical-line {
+    position: absolute;
+    top: 0;
+    left: 100px;
+    width: 2px;
+    height: 100%;
+    background-color: $color-primary;
+    z-index: 15;
 }
